feat: Refactor authentication middleware to use verifySession utility and streamline session verification
