+++ src/utilities/edge-only/index.ts
@@ -1,1 +1,2 @@
 export * from "./google-public-key";
+export * from "./verify-session";
