+++ src/utilities/server-only/index.ts
@@ -1,3 +1,4 @@
 import "server-only";
 
 export * from "./prisma-client";
+export * from "./verify-session";
